Fixed the name error of the cosyvoice model and initially achieved lip synchronization for TTS speech
